Move the new command
