+++ cmd/saucectl/saucectl.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"github.com/saucelabs/saucectl/cli/command/configure"
 	initCmd "github.com/saucelabs/saucectl/cli/command/init"
+	"github.com/saucelabs/saucectl/internal/cmd/new"
 	"github.com/saucelabs/saucectl/internal/cmd/run"
 	"github.com/saucelabs/saucectl/internal/cmd/signup"
 	"github.com/saucelabs/saucectl/internal/setup"
@@ -17,7 +18,6 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/getsentry/sentry-go"
-	"github.com/saucelabs/saucectl/cli/command/new"
 )
 
 var (
